feat: migrate DB to Supabase & restrict phone inputs, show customer phone in requests, save admin notes, and delete custom request post-payment

+++ app/models/audit_log.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, JSON
 from sqlalchemy.sql import func
 from app.database import Base
 
@@ -12,8 +12,8 @@ class AuditLog(Base):
     """
     __tablename__ = "audit_logs"
 
-    id          = Column(Integer, primary_key=True, autoincrement=True)
-    actor_id    = Column(Integer, nullable=True, index=True)
+    id          = Column(BigInteger, primary_key=True, autoincrement=True)
+    actor_id    = Column(BigInteger, nullable=True, index=True)
     actor_email = Column(String, nullable=True)
     action      = Column(String, nullable=False, index=True)
     target_type = Column(String, nullable=True)
